backend/requirements.cloudrun.txt
fastapi
uvicorn[standard]
celery
redis
supabase
python-dotenv
pydantic
pydantic-settings
httpx
alembic
psycopg2-binary
tenacity
pgvector
asyncpg
sqlalchemy
tree_sitter<0.22
tree_sitter_languages
sentence_transformers
google-generativeai
gitpython
